pyproject.toml
[project]
name = "smart-car-lock"
version = "1.0"
description = "open a secure tunnel between your phones to your car allowing to lock and unlock it remotely"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
